feat: 여행 상세 조회 기능 추가 - 완료된 여행 리스트 수정 (폴더 여부에 따라 리스트 구분 반환) Closes #10

+++ triplet/src/main/java/com/ssafy/triplet/travel/service/TravelServiceImpl.java
@@ -3,6 +3,7 @@ package com.ssafy.triplet.travel.service;
 import com.ssafy.triplet.exception.CustomException;
 import com.ssafy.triplet.member.entity.Member;
 import com.ssafy.triplet.member.repository.MemberRepository;
+import com.ssafy.triplet.response.ApiResponse;
 import com.ssafy.triplet.travel.dto.request.TravelRequest;
 import com.ssafy.triplet.travel.dto.response.TravelListResponse;
 import com.ssafy.triplet.travel.dto.response.TravelResponse;
@@ -14,13 +15,16 @@ import com.ssafy.triplet.travel.repository.*;
 import com.ssafy.triplet.travel.util.InviteCodeGenerator;
 import com.ssafy.triplet.travel.util.S3Service;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
 import java.time.LocalDate;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
 
 @Service
@@ -37,22 +41,28 @@ public class TravelServiceImpl implements TravelService {
     @Override
     public TravelResponse createTravel(Long userId, TravelRequest request, MultipartFile image) throws IOException {
         validateTravelRequest(request);
-        Travel travel = buildTravel(userId, request, image);
+        String inviteCode = InviteCodeGenerator.generateInviteCode();
+        Travel travel = buildTravel(userId, request, image, inviteCode);
         Travel savedTravel = travelRepository.save(travel);
         insertTravelMembers(userId, travel.getId());
         saveTravelBudgets(request, savedTravel);
-        return buildTravelResponse(savedTravel, request.getMemberCount());
+        return buildTravelResponse(savedTravel, inviteCode);
     }
 
 //    여행 생성 시 계좌번호 생성
 //    String walletNumber = "124-" + (1000 + new Random().nextInt(9000)) + "-1247";
 
     @Override
-    public TravelResponse updateTravel(Long travelId, TravelRequest request, MultipartFile image) throws IOException {
+    public TravelResponse updateTravel(Long travelId, TravelRequest request, MultipartFile image, Long userId) throws IOException {
         validateTravelRequest(request);
 
         Travel travel = travelRepository.findById(travelId)
-                .orElseThrow(() -> new CustomException("T0005", "해당 여행을 찾을 수 없습니다."));
+                .orElseThrow(() -> new CustomException("T0004", "여행이 존재하지 않습니다."));
+
+        Long creatorId = travelRepository.findCreatorIdByTravelId(travelId);
+        if (!userId.equals(creatorId)) {
+            throw new CustomException("T0011", "여행 생성자가 아닙니다.");
+        }
 
         travel.setStartDate(request.getStartDate());
         travel.setEndDate(request.getEndDate());
@@ -73,13 +83,18 @@ public class TravelServiceImpl implements TravelService {
 
         updateTravelBudgets(request, updatedTravel);
 
-        return buildTravelResponse(updatedTravel, request.getMemberCount());
+        return buildTravelResponse(updatedTravel, travel.getInviteCode());
     }
 
     @Override
-    public void deleteTravel(Long travelId) {
+    public void deleteTravel(Long travelId, Long userId) {
         if (travelRepository.existsById(travelId)) {
-            travelRepository.deleteById(travelId);
+            Long creatorId = travelRepository.findCreatorIdByTravelId(travelId);
+            if (!userId.equals(creatorId)) {
+                throw new CustomException("T0011", "여행 생성자가 아닙니다.");
+            } else {
+                travelRepository.deleteById(travelId);
+            }
         } else {
             throw new CustomException("T0004", "여행이 존재하지 않습니다.");
         }
@@ -91,21 +106,34 @@ public class TravelServiceImpl implements TravelService {
         List<Travel> travelList = travelRepository.findOngoingTravelsByUserId(userId, today);
         List<TravelListResponse> responseList = new ArrayList<>();
         for (Travel travel : travelList) {
-            responseList.add(convertToTravelListResponse(travel));
+            responseList.add(convertToTravelListResponse(travel, userId));
         }
         return responseList;
     }
 
     @Override
-    public List<TravelListResponse> getTravelCompleteList(Long userId) {
+    public Map<String, List<TravelListResponse>> getTravelCompleteList(Long userId) {
         LocalDate today = LocalDate.now();
         List<Travel> travelList = travelRepository.findCompletedTravelsByUserId(userId, today);
-        List<TravelListResponse> responseList = new ArrayList<>();
+
+        List<TravelListResponse> folderlessList = new ArrayList<>();
+        List<TravelListResponse> withFolderList = new ArrayList<>();
         for (Travel travel : travelList) {
-            responseList.add(convertToTravelListResponse(travel));
+            TravelListResponse response = convertToTravelListResponse(travel, userId);
+            Long folderId = travelRepository.findFolderIdByUserId(userId, travel.getId());
+            if (folderId == null) {
+                folderlessList.add(response);
+            } else {
+                withFolderList.add(response);
+            }
         }
-        return responseList;
+        Map<String, List<TravelListResponse>> result = new HashMap<>();
+        result.put("folderlessTravels", folderlessList);
+        result.put("withFolderTravels", withFolderList);
+
+        return result;
     }
+
 
     @Override
     public List<TravelListResponse> getTravelUpcomingList(Long userId) {
@@ -113,9 +141,16 @@ public class TravelServiceImpl implements TravelService {
         List<Travel> travelList = travelRepository.findUpcomingTravelsByUserId(userId, today);
         List<TravelListResponse> responseList = new ArrayList<>();
         for (Travel travel : travelList) {
-            responseList.add(convertToTravelListResponse(travel));
+            responseList.add(convertToTravelListResponse(travel, userId));
         }
         return responseList;
+    }
+
+    @Override
+    public TravelResponse getTravel(Long travelId) {
+        Travel travel = travelRepository.findById(travelId)
+                .orElseThrow(() -> new CustomException("T0004", "여행이 존재하지 않습니다."));
+        return buildTravelResponse(travel, travel.getInviteCode());
     }
 
     // 필수 값 및 날짜 검증 메서드 (여행 생성, 여행 수정)
@@ -137,8 +172,9 @@ public class TravelServiceImpl implements TravelService {
     }
 
     // Travel 객체 생성 메서드 (여행 생성)
-    private Travel buildTravel(Long userId, TravelRequest request, MultipartFile image) throws IOException {
+    private Travel buildTravel(Long userId, TravelRequest request, MultipartFile image, String inviteCode) throws IOException {
         Travel travel = new Travel();
+        travel.setInviteCode(inviteCode);
         travel.setStartDate(request.getStartDate());
         travel.setEndDate(request.getEndDate());
         travel.setTitle(request.getTitle());
@@ -172,17 +208,13 @@ public class TravelServiceImpl implements TravelService {
     }
 
     // 응답 생성 메서드 (여행 생성, 여행 수정)
-    private TravelResponse buildTravelResponse(Travel travel, int memberCount) {
+    private TravelResponse buildTravelResponse(Travel travel, String inviteCode) {
         TravelResponse response = new TravelResponse();
-
-        if (memberCount >= 2) {
-            String inviteCode = InviteCodeGenerator.generateInviteCode();
-            response.setInviteCode(inviteCode);
-        }
-
+        response.setInviteCode(inviteCode);
         response.setTravelId(travel.getId());
         response.setCountry(travel.getCountry().getName());
         response.setCountryId(travel.getCountry().getId());
+        response.setCurrency(travel.getCountry().getCurrency());
         response.setStartDate(travel.getStartDate());
         response.setEndDate(travel.getEndDate());
         response.setTitle(travel.getTitle());
@@ -190,6 +222,9 @@ public class TravelServiceImpl implements TravelService {
         response.setMemberCount(travel.getMemberCount());
         response.setTotalBudget(travel.getTotalBudget());
         response.setCreatorId(travel.getCreatorId());
+        response.setStatus(travel.isStatus());
+        response.setShared(travel.isShared());
+        response.setShareStatus(travel.isShareStatus());
 
         List<TravelBudget> savedBudgets = travelBudgetRepository.findByTravel(travel);
         response.setBudgets(
@@ -236,7 +271,7 @@ public class TravelServiceImpl implements TravelService {
     }
 
     // 여행 리스트
-    private TravelListResponse convertToTravelListResponse(Travel travel) {
+    private TravelListResponse convertToTravelListResponse(Travel travel, Long userId) {
         TravelListResponse response = new TravelListResponse();
         response.setTravelId(travel.getId());
         response.setTitle(travel.getTitle());
@@ -246,8 +281,11 @@ public class TravelServiceImpl implements TravelService {
         response.setCountryId(travel.getCountry().getId());
         response.setCountryName(travel.getCountry().getName());
         response.setMemberCount(travel.getMemberCount());
+        response.setCurrency(travel.getCountry().getCurrency());
         response.setTotalBudget(travel.getTotalBudget());
-
+        Long folderId = travelRepository.findFolderIdByUserId(userId, travel.getId());
+        response.setFolderId(folderId);
+        response.setFolderName(travelRepository.findFolderNameByUserId(folderId));
         return response;
     }
 }
